add convenient functions for solver construction
